fix(authentication): user is able to pass auth middleware when blocked

+++ modules/authentication/src/handlers/common.ts
@@ -40,6 +40,9 @@ export class CommonHandlers implements IAuthenticationStrategy {
     if (!oldRefreshToken.user) {
       throw new GrpcError(status.PERMISSION_DENIED, 'Invalid user');
     }
+    if (!(oldRefreshToken.user as User).active) {
+      throw new GrpcError(status.PERMISSION_DENIED, 'User is blocked');
+    }
 
     // delete the old refresh token
     await RefreshToken.getInstance().deleteOne({ _id: oldRefreshToken._id });
